+++ app/search.js
@@ -66,10 +66,12 @@ const throttling = throttle(async ({ input, results, historyValue = false } = {}
         ajax: true,
     };
 
+    const startEvent = new Event('tntsearch:start');
     const query = Object.keys(params)
         .map(k => `${k}=${params[k]}`)
         .join('&');
 
+    input.dispatchEvent(startEvent);
     fetch(`${data.uri}?${query}`)
         .then((response) => response.text())
         .then((response) => {
@@ -79,8 +81,10 @@ const throttling = throttle(async ({ input, results, historyValue = false } = {}
             return response;
         })
         .then((response) => {
+            const doneEvent = new Event('tntsearch:done');
             results.style.display = '';
             results.innerHTML = response;
+            input.dispatchEvent(doneEvent);
 
             return response;
         });